--- contrib/promexporter/internal/explorer/generate.go
@@ -1,3 +1,0 @@
-package explorer
-
-//go:generate go run github.com/synapsecns/sanguine/contrib/promexporter/internal/explorer/contrib
