+++ contrib/promexporter/internal/exporter.go
@@ -1,0 +1,10 @@
+package internal
+
+import "github.com/synapsecns/sanguine/core/metrics"
+
+type Exporter struct {
+}
+
+func NewExporter(handler metrics.Handler) {
+	//handler.Meter().NewCounter()
+}
