+++ contrib/promexporter/internal/gql/dfk/client.gen.go
