+++ contrib/promexporter/internal/gql/dfk/generate.go
