+++ contrib/promexporter/internal/gql/dfk/queries/queries.graphql
