+++ contrib/promexporter/internal/gql/explorer/generate.go
@@ -1,0 +1,3 @@
+package explorer
+
+//go:generate go run github.com/synapsecns/sanguine/contrib/promexporter/internal/gql/explorer/contrib
